feat: restructure public website navigation (#90)
Reorganize site navigation into new sections: Mosquito Control,
Mosquito Surveillance, and top-level Job Opportunities. Move existing
pages to new URL paths with 301 redirects from old URLs. Add stub
pages for upcoming content.
