+++ .changeset/updated-navigation.md
@@ -1,0 +1,5 @@
+---
+"public": minor
+---
+
+Restructure public website navigation: add Mosquito Control, Mosquito Surveillance, and Job Opportunities sections; move existing pages to new URL paths with 301 redirects from old URLs; add stub pages for spray notice, aerial larviciding, weekly activity, source checklist, and municipal packet.
