+++ apps/public/src/routes/about/how-we-control-mosquitoes.tsx
@@ -1,160 +1,10 @@
-import { createFileRoute, Link } from "@tanstack/react-router";
+import { createFileRoute, redirect } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/about/how-we-control-mosquitoes")({
-	component: RouteComponent,
+	beforeLoad: () => {
+		throw redirect({
+			to: "/mosquito-control/how-we-control-mosquitoes",
+			statusCode: 301,
+		});
+	},
 });
-
-function RouteComponent() {
-	return (
-		<article className="prose lg:prose-xl max-w-none">
-			<h1>How We Control Mosquitoes</h1>
-			<h2>Integrated Mosquito Management</h2>
-			<p>
-				Integrated Mosquito Management is the scientifically based, best
-				management practices that includes mosquito surveillance, public
-				education, source reduction, biological control, and insecticide
-				control. This method of mosquito control has been adopted by the MCMEC
-				Commissioners and is endorsed by the NJ Agricultural Experiment Station
-				(NJAES), Rutgers University, US Environmental Protection Agency (EPA),
-				and the New Jersey Department of Environmental Protection (NJDEP).
-			</p>
-			<h2>Surveillance</h2>
-			<p>
-				All mosquito control actions start with surveillance. The Commission
-				conducts surveillance for larval mosquitoes, adult mosquitoes, and
-				mosquitoes that transmit diseases to humans. The Commission employs
-				different types of mosquito traps to monitor the abundance of the adult
-				mosquito populations and for the presence of disease in those mosquito
-				populations. Mosquito larvae are monitored by visually inspecting
-				stagnant water sources for larvae. Click to view Middlesex County's{" "}
-				<a
-					href="https://drive.google.com/file/d/1WWu4J0ywy2bQu_neVvWwUAOix1Xkigm3/view"
-					rel="noopener noreferrer"
-					target="_blank"
-				>
-					Weekly Mosquito Activity
-				</a>
-				.
-			</p>
-			<h2>Public Education & Outreach</h2>
-			<p>
-				Public education is an essential part of any mosquito control program
-				encourages the public to practice source reduction around their homes
-				and protect themselves from mosquito-borne disease at home and when
-				traveling. The Commission attends large public events as well as
-				provides classroom lectures to schools and other groups.
-			</p>
-			<h2>Source Reduction</h2>
-			<h3>Water Management</h3>
-			<p>
-				In many instances, stormwater facilities such as ditches may get clogged
-				and create a mosquito habitat. The Commission’s Water Management Program
-				may take actions to return the flow to the structure. Project completion
-				may require a large investment of time and other resources. To maximize
-				efforts, much of our work focuses on improving, cleaning, and restoring
-				existing storm water structures to combat some of the prolific mosquito
-				larval habitats in the County. In doing so, the Commission seeks to
-				minimize the need for insecticide applications in these areas and
-				provide long-term protection for residents from nuisance and
-				disease-carrying mosquitoes.
-			</p>
-			<p>
-				By following recommendations in the “Best Management Practices for
-				Mosquito Control and Freshwater Wetlands Management,” all water
-				management projects are done with a minimal amount of temporary
-				disturbance to the surrounding environment. Re-vegetation of disturbed
-				areas usually takes place within a few weeks of work being completed
-				during the growing season. Water management projects are carried out
-				under the authorization of various permits issued by the NJ Department
-				of Environmental Protection (NJDEP). Many water management projects are
-				carried out with the cooperation of other county departments and local
-				municipalities.
-			</p>
-			<h3>Site Plan Reviews</h3>
-			<p>
-				The Commission reviews applications and plans submitted to the Middlesex
-				County Planning Board Development Review Division for any potential
-				mosquito issues. Occasionally, municipalities request Commission
-				approval of proposed stormwater facilities not covered under the County
-				Review Division. The Commission verifies proposed stormwater management
-				systems follow County guidelines as well as the NJDEP's as stated in the
-				NJDEP's Stormwater Best Management Practices manual for any potential
-				issues relating to mosquito production. Upon review of site plans, the
-				Commission forwards a report to the County Planning Board with
-				recommendations for revisions if necessary.
-			</p>
-			<h3>Tire Recycling</h3>
-			<p>
-				Improperly stored or illegally disposed tires can collect water and
-				become habitats for large populations of mosquito that can transmit
-				potentially fatal diseases such as eastern equine encephalitis, West
-				Nile virus, chikungunya, Zika and dengue fever. By removing illegally
-				dumped tires, the Commission staff eliminate mosquito habitat. The
-				Commission has collected and recycled over 100,000 tires from routinely
-				inspected habitats since the beginning of the abatement program.{" "}
-				<a
-					href="https://www.middlesexcountynj.gov/government/departments/department-of-public-safety-and-health/office-of-health-services/solid-waste-management-program"
-					rel="noopener noreferrer"
-					target="_blank"
-				>
-					Click here
-				</a>{" "}
-				for information on how you can recycle your personal scrap tires.
-			</p>
-			<h2>Biological Control</h2>
-			<p>
-				The Commission also stocks mosquito-eating fish in unkempt swimming
-				pools, small ponds and other standing water bodies that may breed
-				mosquitoes. While not eliminating the habitat, biological control can be
-				an inexpensive, efficient, and environmentally friendly means of
-				controlling mosquito populations. Fish are stocked according to
-				guidelines provided by the NJDEP State Mosquito Control Commission
-				(SMCC) and the Office of Mosquito Control Coordination (OMCC).{" "}
-				<a
-					href="https://docs.google.com/forms/d/e/1FAIpQLSe5DMK2IoFxyqKNBmUtn_z-aKdlpry4YzsZp7CC8W_uoYU_Xw/viewform"
-					rel="noopener noreferrer"
-					target="_blank"
-				>
-					Request for fish stocking.
-				</a>
-			</p>
-			<h2>Insecticide Control</h2>
-			<h3>Larval Control</h3>
-			<p>
-				Over 14,000 (potential) larval habitat sites are inspected and serviced
-				on a regular basis by the Mosquito Commission. The Commission's
-				NJDEP-licensed Inspectors regularly check these sites for mosquitoes and
-				treat them with insecticides when mosquitoes are found. Some of these
-				sources are large floodplains and swamps capable of periodically
-				producing a great number of mosquitoes. Due their size and lack of
-				accessibility they are best{" "}
-				<a
-					href="https://www.google.com/maps/d/u/0/viewer?mid=1405E_kiyQ4_fjzdJn9jkFzY7SQE&ll=40.45464790332061%2C-74.32170811513733&z=12"
-					rel="noopener noreferrer"
-					target="_blank"
-				>
-					treated by aircraft.
-				</a>
-			</p>
-			<h3>Adult Mosquito Control</h3>
-			<p>
-				The Commission monitors nuisance and disease-causing mosquitoes
-				throughout the mosquito season using specialized traps. Should an
-				infected mosquito be found or if mosquito populations exceed nuisance
-				thresholds, insecticides (adulticides) are used to control them.
-				Adulticides are applied using truck-mounted sprayers or aerially by
-				helicopter. To find out if an application is planned for your area visit
-				our{" "}
-				<a
-					href="https://drive.google.com/file/d/1P5jI4w5LobA8ztR5BhPr4JTpvJuZTlKf/view"
-					rel="noopener noreferrer"
-					target="_blank"
-				>
-					Spray Schedule
-				</a>{" "}
-				page. Insecticides used by the Commission can be found{" "}
-				<Link to="/about/mosquito-control-products">here</Link>.
-			</p>
-		</article>
-	);
-}
