+++ apps/public/src/routes/mosquito-control/spray-notice.tsx
@@ -1,0 +1,14 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+export const Route = createFileRoute("/mosquito-control/spray-notice")({
+	component: RouteComponent,
+});
+
+function RouteComponent() {
+	return (
+		<article className="prose lg:prose-xl max-w-none">
+			<h1>Public Notice for Adult Mosquito Control Treatment</h1>
+			<p>This page is under construction. Please check back soon.</p>
+		</article>
+	);
+}
